test(workflows): migrate resume-picker integration harness to the host session-picker capability
The /workflow resume integration suites still drove the picker through the
deleted remote-rendered ctx.ui.custom path and failed in CI (the pre-commit
gate only runs test:unit). buildPrintCtxWithRealCustom now exposes the real
in-process capability: ctx.ui.hostSessionPicker is openLocalHostSessionPicker
over a promise-returning ui.custom seam that captures into the same
CapturedCustomCall array, so customCalls[0].component is the REAL
SessionSelectorComponent and the existing chrome/row/selection/escape
assertions drive the production mount path unchanged. Run/stage store
fixtures move to overlay-entrypoints-run-fixtures.ts (re-exported) to stay
under the 500-line file gate.

Assistant-model: Claude Fable 5
